platform: msm_shared: Fix cache flush issue
Flush the cache before handing over the data to storage driver. Earlier
assumption was all the data to storage comes from USB. As USB driver already
flushes the data, flushing again for mmc driver was not needed. But there are
other instances where other drivers would call into storage drivers for writing
data to storage and we need to flush the cache

Change-Id: I4698729768e3ea5ba3c0e2f735b5941481f8b0b3